switch all sm-abi usages of int for uintptr_t, using int causes various bad behaviors with sign extension

+++ sm-sbi.h
@@ -3,9 +3,9 @@
 
 #include <stdint.h>
 #include <stddef.h>
-int mcall_sm_create_enclave(unsigned long base, unsigned long size);
-int mcall_sm_destroy_enclave(unsigned long eid);
-int mcall_sm_run_enclave(unsigned long eid, unsigned long ptr);
-int mcall_sm_exit_enclave(unsigned long ret);
-int mcall_sm_not_implemented(unsigned long a0);
+uintptr_t mcall_sm_create_enclave(unsigned long base, unsigned long size);
+uintptr_t mcall_sm_destroy_enclave(unsigned long eid);
+uintptr_t mcall_sm_run_enclave(unsigned long eid, unsigned long ptr);
+uintptr_t mcall_sm_exit_enclave(unsigned long ret);
+uintptr_t mcall_sm_not_implemented(unsigned long a0);
 #endif
